# src/pwa/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebAZ Protocol API",
    "version": "0.4.14",
    "description": "Auto-generated endpoint inventory (640 endpoints, 17 with full schema). See docs/ for design context."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "api_key"
      },
      "webauthnToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WebAuthn-Token"
      }
    }
  },
  "paths": {
...
    "/api/disputes/cases/:case_id/comment": {
      "post": {
        "summary": "写评论 — 当事人禁评，一人一案一次",
        "tags": [
          "dispute"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/disputes/cases/:case_id/comments/:comment_id/reply": {
      "post": {
        "summary": "W5 子回复 — 任意人可对顶层评论回复多次（不受\"一人一案一次\"限制）",
        "tags": [
          "dispute"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/disputes/cases/:case_id/fairness": {
      "post": {
        "summary": "公正度投票（👍 / 👎）— 一人一案一票",
        "tags": [
          "dispute"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}